% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{find_bam_files}
\alias{find_bam_files}
\title{find_bam_files}
\usage{
find_bam_files(wd)
}
\arguments{
\item{wd}{}
}
\value{

}
\description{
find_bam_files
}
\examples{
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
bam_files = find_bam_files(wd)[1:10]
bam_files
}
