% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{find_SJ.out.tab_files}
\alias{find_SJ.out.tab_files}
\title{find_SJ.out.tab_files}
\usage{
find_SJ.out.tab_files(wd)
}
\arguments{
\item{wd}{}
}
\value{

}
\description{
find_SJ.out.tab_files
}
\examples{
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)[1:10]
sj_files
}
