% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_splicing_from_bam_files}
\alias{load_splicing_from_bam_files}
\title{load_splicing_from_bam_files}
\usage{
load_splicing_from_bam_files(
  bam_files,
  view_gr,
  trim_extension = ".Aligned.sortedByCoord.out.bam"
)
}
\arguments{
\item{bam_files}{}
}
\value{

}
\description{
load_splicing_from_bam_files
}
\examples{
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)[1:10]

features = load_gtf("~/gencode.v36.annotation.gtf", gene_of_interest = "IKZF1")

load_splicing_from_SJ.out.tab_files(sj_files, features$view_gr)
}
