% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_spanning_reads.R
\name{plot_sj_pileups}
\alias{plot_sj_pileups}
\title{Title}
\usage{
plot_sj_pileups(sj_dt, bam_files, view_gr, max_span = 200)
}
\arguments{
\item{max_span}{}
}
\value{

}
\description{
Title
}
\examples{
options(mc.cores = 20)
set.seed(0)

# wd = "/slipstream/home/joeboyd/R/SF_Ikaros_splicing/data"
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)
bam_files = find_bam_files(wd)
features = load_gtf("~/gencode.v36.annotation.gtf", gene_of_interest = "IKZF1")

sp_sj_dt = load_splicing_from_SJ.out.tab_files(sj_files, view_gr = features$view_gr)
# sp_bam_dt = load_splicing_from_bam_files(bam_files, view_gr = features$view_gr)
sp_sj_dt.sel = sp_sj_dt[number_unique > 400]

plots = plot_sj_pileups(sp_sj_dt.sel, bam_files[1:4], features$view_gr)
plots$assembled
}
