% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_view_pileup_and_splicing}
\alias{plot_view_pileup_and_splicing}
\title{plot_view_pileup_and_splicing}
\usage{
plot_view_pileup_and_splicing(
  splice_dt,
  bam_files,
  view_gr,
  plot_title = "",
  splice_height_var = "number_unique",
  return_data = FALSE,
  anno_dt = NULL,
  pool_by = NULL,
  pool_FUN = sum,
  sample_sep = "\\n",
  exons_to_show = NULL,
  win_size = 50,
  prof_dt = NULL,
  prof_dt.add = NULL
)
}
\arguments{
\item{prof_dt.add}{}
}
\value{

}
\description{
plot_view_pileup_and_splicing
}
\examples{
options(mc.cores = 20)
set.seed(0)

# wd = "/slipstream/home/joeboyd/R/SF_Ikaros_splicing/data"
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)[1:10]
bam_files = find_bam_files(wd)[1:10]
features = load_gtf("~/gencode.v36.annotation.gtf", gene_of_interest = "IKZF1")

sp_sj_dt = load_splicing_from_SJ.out.tab_files(sj_files, view_gr = features$view_gr)
sp_sj_dt.sel = sp_sj_dt[number_unique > 5]

plots = plot_view_pileup_and_splicing(sp_sj_dt.sel, bam_files[1:4], features$view_gr, exons_to_show = features$goi_exon_dt[transcript_id == "ENST00000331340.8"])

bam_qdt = data.table(file = bam_files)
bam_qdt[, sample := sub(".Aligned.sorted.+", "", basename(file))]

anno_dt = unique(sp_sj_dt.sel[, .(sample)])
anno_dt$group = rep(c("a", "b"), each = 5)
anno_dt$group2 = rep(c("c", "d"), 5)
debug(plot_view_pileup_and_splicing)

plots = plot_view_pileup_and_splicing(sp_sj_dt.sel, bam_files[1:4], features$view_gr,
  exons_to_show = features$goi_exon_dt[transcript_id == "ENST00000331340.8"])

plots.pooled = plot_view_pileup_and_splicing(sp_sj_dt.sel, bam_qdt[1:4,], features$view_gr,
  anno_dt = anno_dt,
  pool_by = c("group", "group2"),
  exons_to_show = features$goi_exon_dt[transcript_id == "ENST00000331340.8"])

cowplot::plot_grid(plots, plots.pooled)
}
