% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSplice.R
\name{suppa_clusterEvents}
\alias{suppa_clusterEvents}
\title{Title}
\usage{
suppa_clusterEvents(wd, PSI_todo = unlist(SPLICE_EVENTS), output_location = wd)
}
\arguments{
\item{output_location}{}
}
\value{

}
\description{
Title
}
\examples{
wd = "suppa2_diff"
PSI_todo = unlist(SSV_SPLICE_EVENTS)
output_location = wd
psi = PSI_todo[1]
suppa_clusterEvents(wd)
}
