% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSplice.R
\name{suppa_diffSplice}
\alias{suppa_diffSplice}
\title{Title}
\usage{
suppa_diffSplice(
  wd,
  ref_location = "~/indexes/MM10/SUPPA2",
  PSI_todo = unlist(SPLICE_EVENTS),
  output_location = wd
)
}
\arguments{
\item{PSI_todo}{}
}
\value{

}
\description{
Title
}
\examples{

wd = "suppa2_diff"
tpm_files = dir(wd, pattern = ".tpm$", full.names = TRUE)
PSI_todo = unlist(SPLICE_EVENTS)
psi = PSI_todo[[1]]
psi_files = dir(wd, pattern = paste0(psi, ".psi$"), full.names = TRUE)
ref_location = "~/indexes/HG38/SUPPA2"
suppa_diffSplice(wd, ref_location)
}
