% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSplice.R
\name{suppa_joinFiles}
\alias{suppa_joinFiles}
\title{suppa_joinFiles}
\usage{
suppa_joinFiles(
  input_files,
  output_name,
  run_TPM = TRUE,
  PSI_todo = unlist(SPLICE_EVENTS),
  bam_suffix = ".Aligned.+",
  output_location = "suppa2_diff"
)
}
\arguments{
\item{output_location}{}
}
\value{

}
\description{
suppa_joinFiles
}
\examples{
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
bam_files = find_bam_files(wd)
input_files = bam_files[1:5]
suppa_joinFiles(input_files, "test")

suppa_joinFiles(bam_files[1:5], "test1")
suppa_joinFiles(bam_files[6:10], "test2")
}
