% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_splicing_from_bam_files}
\alias{load_splicing_from_bam_files}
\title{Title}
\usage{
load_splicing_from_bam_files(
  bam_files,
  view_gr,
  trim_extension = ".Aligned.sortedByCoord.out.bam"
)
}
\arguments{
\item{bam_files}{}
}
\value{

}
\description{
Title
}
