% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_splicing_from_SJ.out.tab_files}
\alias{load_splicing_from_SJ.out.tab_files}
\title{Title}
\usage{
load_splicing_from_SJ.out.tab_files(sj_files, view_gr)
}
\arguments{
\item{view_gr}{}
}
\value{

}
\description{
Title
}
