% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sj_pileups}
\alias{plot_sj_pileups}
\title{plot_sj_pileups}
\usage{
plot_sj_pileups(
  sj_dt,
  bam_files,
  view_gr,
  max_span = 200,
  pool_by = NULL,
  pool_FUN = sum,
  sample_sep = "\\n",
  prof_at_starts = NULL,
  prof_at_ends = NULL,
  rel_heights.assembly = c(2, 1)
)
}
\arguments{
\item{prof_at_ends}{}
}
\value{

}
\description{
plot_sj_pileups
}
\examples{
options(mc.cores = 20)
set.seed(0)

# wd = "/slipstream/home/joeboyd/R/SF_Ikaros_splicing/data"
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)[1:10]
bam_files = find_bam_files(wd)[1:10]
features = load_gtf("~/gencode.v36.annotation.gtf", gene_of_interest = "IKZF1")

sp_sj_dt = load_splicing_from_SJ.out.tab_files(sj_files, view_gr = features$view_gr)
# sp_bam_dt = load_splicing_from_bam_files(bam_files, view_gr = features$view_gr)
sp_sj_dt.sel = sp_sj_dt[number_unique > 100]

plots = plot_sj_pileups(sp_sj_dt.sel, bam_files[1:4], features$view_gr)
plots$assembled

bam_qdt = data.table(file = bam_files[1:10], a = LETTERS[rep(c(1, 2), each = 5)], b = LETTERS[rep(c(3, 4), 5)])
plots.pooled = plot_sj_pileups(sp_sj_dt.sel, bam_qdt, features$view_gr, pool_by = c("a", "b"))
}
