% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_splice_heatmap}
\alias{plot_splice_heatmap}
\title{Title}
\usage{
plot_splice_heatmap(
  dtw,
  sel_sample,
  pdf_name,
  name,
  min_test_value = 20,
  win_size = 1
)
}
\arguments{
\item{win_size}{}
}
\value{

}
\description{
Title
}
