% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_view_pileup_and_splicing}
\alias{plot_view_pileup_and_splicing}
\title{plot_view_pileup_and_splicing}
\usage{
plot_view_pileup_and_splicing(
  splice_dt,
  bam_files,
  view_gr,
  return_data = FALSE,
  pool_by = NULL,
  pool_FUN = sum,
  sample_sep = "\\n"
)
}
\arguments{
\item{sample_sep}{}
}
\value{

}
\description{
plot_view_pileup_and_splicing
}
\examples{
options(mc.cores = 20)
set.seed(0)

# wd = "/slipstream/home/joeboyd/R/SF_Ikaros_splicing/data"
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
sj_files = find_SJ.out.tab_files(wd)[1:10]
bam_files = find_bam_files(wd)[1:10]
features = load_gtf("~/gencode.v36.annotation.gtf", gene_of_interest = "IKZF1")

sp_sj_dt = load_splicing_from_SJ.out.tab_files(sj_files, view_gr = features$view_gr)
sp_sj_dt.sel = sp_sj_dt[number_unique > 100]

plots = plot_view_pileup_and_splicing(sp_sj_dt.sel, bam_files[1:4], features$view_gr)
}
