% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSplice.R
\name{suppa_clusterEvents}
\alias{suppa_clusterEvents}
\title{suppa_clusterEvents}
\usage{
suppa_clusterEvents(wd, PSI_todo = unlist(SPLICE_EVENTS), output_location = wd)
}
\arguments{
\item{output_location}{}
}
\value{

}
\description{
suppa_clusterEvents
}
\examples{
wd = "suppa2_diff"
suppa_clusterEvents(wd)
}
