% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSplice.R
\name{suppa_diffSplice}
\alias{suppa_diffSplice}
\title{suppa_diffSplice}
\usage{
suppa_diffSplice(
  wd,
  ref_location = "~/indexes/MM10/SUPPA2",
  PSI_todo = unlist(SPLICE_EVENTS),
  output_location = wd
)
}
\arguments{
\item{PSI_todo}{}
}
\value{

}
\description{
suppa_diffSplice
}
\examples{

wd = "suppa2_diff"
ref_location = "~/indexes/HG38/SUPPA2"
suppa_diffSplice(wd, ref_location)
}
